Chore: remove temps_linux
