--- src/detection/temps/temps_linux.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "fastfetch.h"
-
-double ffDetectCPUTemp(void);
